requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blis==1.0.2
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
cryptography==43.0.3
cymem==2.0.10
DateTime==5.5
distro==1.9.0
docx2txt==0.8
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
fastapi==0.115.5
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.5
idna==3.10
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.11
networkx==3.4.2
nltk==3.9.1
numpy==2.0.2
ollama==0.4.4
openai==1.57.3
packaging==24.2
pandas==2.2.3
pdfminer.six==20240706
plac==1.1.3
preshed==3.0.9
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pyresparser==1.0.6
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
safetensors==0.4.5
shellingham==1.5.4
six==1.17.0
smart-open==7.0.5
sniffio==1.3.1
sortedcontainers==2.4.0
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.0
starlette==0.41.3
sympy==1.13.1
thinc==8.3.2
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.47.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.0
zope.interface==7.2
